+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 # Change Log
 
+### *(2015-03-01)*
+ * Better button's color on API < 21
+ * Add compatibility to API >= 10
+
+--------------
+
 ### *(2015-02-28)*
  * Creat RippleButton